+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# tidyselect (development version)
+# tidyselect 1.1.0
 
 * Predicate functions must now be wrapped with `where()`.